# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Hello World App
# during dev, should point to your local machine
host: localhost:9000
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
securityDefinitions:
    Bearer:
        type: apiKey
        name: Authorization
        in: header
paths:
  /userreg:
    # binds a127 app logic to a route
    x-swagger-router-controller: user
    post:
      description: Returns 'Signup' to the caller
      # used as the method name of the controller
      operationId: userreg
      parameters:
        - name: "body"
          in: "body"
          description: The email of the person to whom to say hello
          required: false
          schema:
            $ref: "#/definitions/signup" 
          
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /userlogin:
    # binds a127 app logic to a route
    x-swagger-router-controller: user
    post:
      description: Returns 'Login' to the caller
      # used as the method name of the controller
      operationId: userlogin
      parameters:
        - name: "body"
          in: "body"
          description: The email of the person to whom to say hello
          required: false
          schema:
            $ref: "#/definitions/login" 
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"            
...
  /get-my-profile:
    # binds a127 app logic to a route
    x-swagger-router-controller: user
    get:
      description: Returns 'get_task' to the caller
      # used as the method name of the controller
      operationId: get_my_profile
      security:
        - Bearer: []
      parameters:
        - name: userid
          in: query
          description: The email of the person to whom to say hello
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        "403":
          description: "Access Denied"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse" 
  /update-profile:
    # binds a127 app logic to a route
    x-swagger-router-controller: user
    put:
      description: Returns 'Timesheet' to the caller
      # used as the method name of the controller
      operationId: updateprofile
      security:
        - Bearer: []
      parameters:
        - name: "body"
          in: "body"
          description: The email of the person to whom to say hello
          required: false
          schema:
            $ref: "#/definitions/update-profile"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        "403":
          description: "Access Denied"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"            
...
  /fullTimesheetWeekly:
    # binds a127 app logic to a route
    x-swagger-router-controller: weekly_timesheet
    get:
      description: Returns 'rpttimecard' to the caller
      # used as the method name of the controller
      operationId: fullTimesheetWeekly
      security:
        - Bearer: []
      parameters:
        - name: userID
          in: query
          description: The email of the person to whom to say hello
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        "403":
          description: "Access Denied"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /fullTimesheetAdminWeekly:
    # binds a127 app logic to a route
    x-swagger-router-controller: weekly_timesheet
    get:
      description: Returns 'rpttimecard' to the caller
      # used as the method name of the controller
      operationId: fullTimesheetAdminWeekly
      security:
        - Bearer: []
      parameters:
        - name: userID
          in: query
          description: The email of the person to whom to say hello
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/HelloWorldResponse"
        "403":
          description: "Access Denied"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
  HelloWorldResponse:
    required:
      - message
    properties:
      message:
        type: string
  signup:
    type: "object"
    required:
      - "userrole_id"
      - "first_name"
      - "last_name"
      - "email"
      - "password"
    properties:
      userrole_id:
        type: "string"
        example: "hiring"
      first_name:
        type: "string"
        example: "first_name"
      last_name:
        type: "string"
        example: "last_name"
      email:
        type: "string"
        example: "email@gmail.com"
      password:
        type: "string"
        example: "12345678"
  login:
    type: "object"
    required:
    - "email"
    - "password"
    properties:
      email:
        type: "string"
        example: "email@gmail.com"
      password:
        type: "string"
        example: "12345678"
...
  update-profile:
    type: "object"
...
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string